feat: Refactor homepage styling and add links to nav bar
This commit refactors the styling of the home page to use CSS in more idiomatic
ways (e.g. use flexbox instead of absolute positioning).

It ignores the FAQ accordions for now, since I think others may be actively
working on the functionality for that.

It also crops the background image to a more reasonable size/aspect ratio, so
we don't have to use negative pixels to position it correctly.

Hope this doesn't cause too many merge conflicts!