refactor: Move budget amounts to shared Vuex state
This commit moves the user-selected budget amounts for the balance-budget page
into the Vuex `budget` module. This allows it to be shared across components,
in this case both the overview page as well as the carousel components.

This commit also does a small amount of refactoring to better support this
pattern, for example:
- filtering zero amounts out of the pie chart
- allowing carousel sliders to exceed the real budget (as the overview page
  already does)
- commenting out the individual slider validations for now (didn't have the
  brain power to make it work for now -- may require separate validation for
  each slider to make it easier)