feat: Adding full SF/Oakland budget descriptions
This change also refactors how we handle the department screens in
the walkthough, moving this state to the global store. This allows
us to better mutate and read it from the various components, thereby
cleaning up some of the code and relying on Vue's directives rather
than switching display styles.